requirements.txt
binance_connector_python==0.5.0
pandas==1.2.2
